+++ plugins/scaffolder-actions/scaffolder-backend-module-utils/src/actions/jsonata/yaml.test.ts
@@ -18,6 +18,7 @@ import { createYamlJSONataTransformAction } from './yaml';
 import mock from 'mock-fs';
 import YAML from 'yaml';
 import { getVoidLogger } from '@backstage/backend-common';
+import { Scalar } from 'yaml';
 
 describe('roadiehq:utils:jsonata:yaml:transform', () => {
   const mockContext = {
@@ -90,17 +91,23 @@ describe('roadiehq:utils:jsonata:yaml:transform', () => {
     });
 
     const opts = {
-      indent: 3,
-      noArrayIndent: true,
-      skipInvalid: true,
-      flowLevel: 23,
-      sortKeys: true,
+      blockQuote: true,
+      collectionStyle: 'block' as 'any' | 'block' | 'flow',
+      defaultKeyType: 'PLAIN' as Scalar.Type,
+      defaultStringType: 'BLOCK_LITERAL' as Scalar.Type,
+      directives: false,
+      doubleQuotedAsJSON: false,
+      doubleQuotedMinMultiLineLength: 40,
+      falseStr: 'false',
+      flowCollectionPadding: false,
+      indent: 4,
+      indentSeq: false,
       lineWidth: -1,
-      noRefs: true,
-      noCompatMode: true,
-      condenseFlow: true,
-      quotingType: '"' as const,
-      forceQuotes: true,
+      minContentWidth: 20,
+      nullStr: 'null',
+      simpleKeys: false,
+      singleQuote: false,
+      trueStr: 'true',
     };
 
     await action.handler({