+++ plugins/backend/catalog-backend-module-okta/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @roadiehq/catalog-backend-module-okta
 
+## 0.10.2
+
+### Patch Changes
+
+- d6ae6e9: Release all packages to rollout new metadata
+
+## 0.10.1
+
+### Patch Changes
+
+- 370905e: Improved logging of the orgEntityProvider
+
 ## 0.10.0
 
 ### Minor Changes