+++ plugins/backend/rag-ai-backend-embeddings-aws/CHANGELOG.md
@@ -1,5 +1,27 @@
 # @roadiehq/rag-ai-backend-embeddings-aws
 
+## 0.2.8
+
+### Patch Changes
+
+- d6ae6e9: Release all packages to rollout new metadata
+- Updated dependencies [d6ae6e9]
+  - @roadiehq/rag-ai-backend-retrieval-augmenter@0.3.8
+  - @roadiehq/rag-ai-backend@1.0.2
+  - @roadiehq/rag-ai-node@0.1.6
+
+## 0.2.7
+
+### Patch Changes
+
+- b7c834e: Updated @langchain dependencies
+- Updated dependencies [b7c834e]
+- Updated dependencies [caa7d57]
+- Updated dependencies [b7c834e]
+  - @roadiehq/rag-ai-backend-retrieval-augmenter@0.3.7
+  - @roadiehq/rag-ai-backend@1.0.0
+  - @roadiehq/rag-ai-node@0.1.5
+
 ## 0.2.6
 
 ### Patch Changes