+++ plugins/backend/rag-ai-backend-retrieval-augmenter/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @roadiehq/rag-ai-backend-retrieval-augmenter
 
+## 0.3.8
+
+### Patch Changes
+
+- d6ae6e9: Release all packages to rollout new metadata
+- Updated dependencies [d6ae6e9]
+  - @roadiehq/rag-ai-node@0.1.6
+
+## 0.3.7
+
+### Patch Changes
+
+- b7c834e: Updated @langchain dependencies
+- Updated dependencies [b7c834e]
+  - @roadiehq/rag-ai-node@0.1.5
+
 ## 0.3.6
 
 ### Patch Changes