+++ plugins/backend/rag-ai-node/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @roadiehq/rag-ai-node
 
+## 0.1.6
+
+### Patch Changes
+
+- d6ae6e9: Release all packages to rollout new metadata
+
+## 0.1.5
+
+### Patch Changes
+
+- b7c834e: Updated @langchain dependencies
+
 ## 0.1.4
 
 ### Patch Changes