+++ plugins/backend/rag-ai-storage-pgvector/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @roadiehq/rag-ai-storage-pgvector
 
+## 0.1.6
+
+### Patch Changes
+
+- d6ae6e9: Release all packages to rollout new metadata
+- Updated dependencies [d6ae6e9]
+  - @roadiehq/rag-ai-node@0.1.6
+
+## 0.1.5
+
+### Patch Changes
+
+- b7c834e: Updated @langchain dependencies
+- Updated dependencies [b7c834e]
+  - @roadiehq/rag-ai-node@0.1.5
+
 ## 0.1.4
 
 ### Patch Changes