+++ plugins/frontend/backstage-plugin-argo-cd/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @roadiehq/backstage-plugin-argo-cd
 
+## 2.7.0
+
+### Minor Changes
+
+- 971f076: Adds support for Backstage's new frontend system, available via the `/alpha` sub-path export.
+
+## 2.6.7
+
+### Patch Changes
+
+- d6ae6e9: Release all packages to rollout new metadata
+
+## 2.6.6
+
+### Patch Changes
+
+- e816d1e: Adding the terminateOperation flag to the /sync endpoint. It is an optional boolean flag that can be set to true and it will terminate the previous Sync in progress before starting a new sync. This is a non-breaking change because if the flag is not provided and set to true the existing logic will not be impacted.
+
 ## 2.6.5
 
 ### Patch Changes