+++ plugins/frontend/backstage-plugin-github-insights/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @roadiehq/backstage-plugin-github-insights
 
+## 2.4.1
+
+### Patch Changes
+
+- d6ae6e9: Release all packages to rollout new metadata
+
+## 2.4.0
+
+### Minor Changes
+
+- 8d9257d: Fix generating paths for relative markdown links from README files
+
 ## 2.3.31
 
 ### Patch Changes