+++ plugins/frontend/rag-ai/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @roadiehq/rag-ai
 
+## 1.1.1
+
+### Patch Changes
+
+- d6ae6e9: Release all packages to rollout new metadata
+
+## 1.1.0
+
+### Minor Changes
+
+- b7c834e: **BREAKING** Added support for streaming responses
+
+## 1.0.0
+
+### Major Changes
+
+- 01f842c: Add support for react 18, and release version 1.
+
 ## 0.2.7
 
 ### Patch Changes