+++ plugins/scaffolder-actions/scaffolder-backend-module-utils/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @roadiehq/scaffolder-backend-module-utils
 
+## 2.0.1
+
+### Patch Changes
+
+- d6ae6e9: Release all packages to rollout new metadata
+
+## 2.0.0
+
+### Major Changes
+
+- ca608a7: Update yaml options schema for 'yaml' package
+
 ## 1.17.1
 
 ### Patch Changes